Part 5 and 6 complete
Part 5 was complete in last commit but was labelled WIP. The fix was to
recompile the system gschemas, nothing to do with source code.

Part 6 was super easy to implement since it followed the same steps as
previous ui changes. Good workflow!

The transitions are cool but the font picker is slooooooooww.